+++ app/src/main/java/com/example/countryapp/model/util/EntityMapper.kt
@@ -1,0 +1,5 @@
+package com.example.countryapp.model.util
+
+interface EntityMapper<Entity, Model > {
+    fun mapFromEntityToModel (entity: Entity): Model
+}
